Move replace and replace_dict to algorithms sub-package
