+++ claripy/algorithm/__init__.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from .bool_check import is_false, is_true
 from .ite_relocation import burrow_ite, excavate_ite
+from .replace import replace, replace_dict
 from .simplify import simplify
 
 __all__ = (
@@ -9,5 +10,7 @@ __all__ = (
     "excavate_ite",
     "is_false",
     "is_true",
+    "replace",
+    "replace_dict",
     "simplify",
 )
