+++ claripy/ast/base.py
@@ -17,11 +17,11 @@ from typing_extensions import Self
 
 import claripy
 from claripy import operations
-from claripy.errors import BackendError, ClaripyOperationError, ClaripyReplacementError
+from claripy.errors import BackendError, ClaripyOperationError
 from claripy.fp import FSort
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterable, Iterator
+    from collections.abc import Iterable, Iterator
 
     from claripy.annotation import Annotation
     from claripy.backends import Backend
@@ -927,89 +927,6 @@ class Base:
 
         return True
 
-    def replace_dict(
-        self, replacements, variable_set: set[str] | None = None, leaf_operation: Callable[[Base], Base] = lambda x: x
-    ) -> Self:
-        """
-        Returns this AST with subexpressions replaced by those that can be found in `replacements`
-        dict.
-
-        :param variable_set:        For optimization, ast's without these variables are not checked
-                                        for replacing.
-        :param replacements:        A dictionary of hashes to their replacements.
-        :param leaf_operation:      An operation that should be applied to the leaf nodes.
-        :returns:                   An AST with all instances of ast's in replacements.
-        """
-        if variable_set is None:
-            variable_set = set()
-
-        arg_queue = [iter([self])]
-        rep_queue = []
-        ast_queue = []
-
-        while arg_queue:
-            try:
-                ast = next(arg_queue[-1])
-                repl = ast
-
-                if not isinstance(ast, Base):
-                    rep_queue.append(repl)
-                    continue
-
-                if ast.hash() in replacements:
-                    repl = replacements[ast.hash()]
-
-                elif ast.variables >= variable_set:
-                    if ast.is_leaf():
-                        repl = leaf_operation(ast)
-                        if repl is not ast:
-                            replacements[ast.hash()] = repl
-
-                    elif ast.depth > 1:
-                        arg_queue.append(iter(ast.args))
-                        ast_queue.append(ast)
-                        continue
-
-                rep_queue.append(repl)
-
-            except StopIteration:
-                arg_queue.pop()
-
-                if ast_queue:
-                    ast = ast_queue.pop()
-                    repl = ast
-
-                    args = rep_queue[-len(ast.args) :]
-                    del rep_queue[-len(ast.args) :]
-
-                    # Check if replacement occurred.
-                    if any((a is not b for a, b in zip(ast.args, args, strict=False))):
-                        repl = ast.make_like(ast.op, tuple(args))
-                        replacements[ast.hash()] = repl
-
-                    rep_queue.append(repl)
-
-        assert len(arg_queue) == 0, "arg_queue is not empty"
-        assert len(ast_queue) == 0, "ast_queue is not empty"
-        assert len(rep_queue) == 1, ("rep_queue has unexpected length", len(rep_queue))
-
-        return rep_queue.pop()
-
-    def replace(self, old: T, new: T) -> Self:
-        """
-        Returns this AST but with the AST 'old' replaced with AST 'new' in its subexpressions.
-        """
-        self._check_replaceability(old, new)
-        replacements = {old.hash(): new}
-        return self.replace_dict(replacements, variable_set=old.variables)
-
-    @staticmethod
-    def _check_replaceability(old: T, new: T) -> None:
-        if not isinstance(old, Base) or not isinstance(new, Base):
-            raise ClaripyReplacementError("replacements must be AST nodes")
-        if type(old) is not type(new):
-            raise ClaripyReplacementError(f"cannot replace type {type(old)} ast with type {type(new)} ast")
-
     def canonicalize(self, var_map=None, counter=None) -> Self:
         counter = itertools.count() if counter is None else counter
         var_map = {} if var_map is None else var_map
@@ -1019,7 +936,7 @@ class Base:
                 new_name = "canonical_%d" % next(counter)
                 var_map[v.hash()] = v._rename(new_name)
 
-        return var_map, counter, self.replace_dict(var_map)
+        return var_map, counter, claripy.replace_dict(self, var_map)
 
     #
     # these are convenience operations
